Factor out API calls as hooks
